20240509 가봤다고 수정

+++ BE/gaxga/src/main/java/com/teamproject/gaxga/dto/GabowatdagoForm.java
@@ -24,7 +24,7 @@ public class GabowatdagoForm {
     private String lat;
     private String lng;
     private User userCode;
-    private Long likeCount;
+    private Integer likeCount;
